Move connection indicator to its own view and controller
