+++ src/main/resources/edu/wpi/first/tableviewer/ConnectionIndicator.fxml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.StackPane?>
+<StackPane xmlns="http://javafx.com/javafx"
+           xmlns:fx="http://javafx.com/fxml"
+           fx:controller="edu.wpi.first.tableviewer.ConnectionIndicatorController"
+           fx:id="root"
+           styleClass="connection-background"
+           prefHeight="50">
+    <Label fx:id="connectionLabel" text="Something went terribly wrong" styleClass="connection-label"/>
+</StackPane>
